refactor: move settings into native module (#369)
