+++ cmd/internal/serverapp/configuration_composition.go
@@ -1,10 +1,11 @@
 package serverapp
 
 import (
-	"github.com/markhuangai/dense-mem/internal/repository"
 	"github.com/markhuangai/dense-mem/internal/service"
+	settings "github.com/markhuangai/dense-mem/internal/settings"
+	settingscontract "github.com/markhuangai/dense-mem/internal/settings/contract"
 )
 
-func buildConfigurationApplication(repo repository.AppConfigRepository, audit service.AuditService) *service.AppConfigServiceImpl {
-	return service.NewAppConfigService(repo, audit)
+func buildConfigurationApplication(repo settingscontract.AppConfigRepository, audit service.AuditService) *settings.AppConfigServiceImpl {
+	return settings.NewAppConfigService(repo, audit)
 }
