+++ internal/settings/app_config_service_test.go
@@ -1,4 +1,4 @@
-package service
+package settings
 
 import (
 	"context"
